+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added CRUD for users.
+- Added security layer for hasing password when storing a new user.
 
 ## [0.2.1] - 2023-04-30